🛑 Simplify_junc uses navGridHM.

+++ Tools/NavGridPP/functions/NavGridPP/fn_NG_simplify_junc.sqf
@@ -1,39 +1,30 @@
 /*
 Maintainer: Caleb Serafin
+    _navRoadHM  is modified by reference!
     All nearby junctions will be merged into the most centred one.
 
 Arguments:
-    <ARRAY<             navGrid:
-        <OBJECT>            Road
-        <ARRAY<OBJECT>>         Connected roads.
-        <ARRAY<SCALAR>>         True driving distance in meters to connected roads.
-    >>
+    <navRoadHM> _navRoadHM is modified by reference
     <SCALAR> Junctions are only merged if within this distance. (Default = 15)
 
 Return Value:
-    <ARRAY> Simplified navGrid
+    <ARRAY> Same Simplified _navRoadHM reference
 
 Scope: Any, Global Arguments
 Environment: Scheduled
 Public: No
 
 Example:
-    _navGrid = [_navGrid,35] call A3A_fnc_NG_simplify_junc;
+    [_navRoadHM,35] call A3A_fnc_NG_simplify_junc;
 */
 params [
-    ["_navGrid",[],[ [] ]], //ARRAY<  Road, ARRAY<connectedRoad>>, ARRAY<distance>  >
+    "_navRoadHM",
     ["_juncMergeDistance",15,[ 0 ]] // Junctions are only merged if within this distance.
 ];
-private _navGridSimple = +_navGrid;
-
-private _roadIndexNS = createHashMap;
-{
-    _roadIndexNS set [str (_x#0),_forEachIndex];
-} forEach _navGridSimple; // _x is road struct <road,ARRAY<connections>,ARRAY<indices>>
 
 private _diag_step_sub = "";
 
-private _fnc_diag_render = { // call _fnc_diag_render;
+private _fnc_diag_render = {
     [
         "Nav Grid++",
         "<t align='left'>" +
@@ -79,12 +70,8 @@ private _fnc_connectStructs = {
     (_otherStruct#2) pushBack _distance;
 };
 
-call _fnc_diag_render;
-
-private _orphanedIndices = [];
-
 private _fnc_consumeStruct = {
-    params ["_myStruct","_otherStruct"]; // _otherStruct is consumed & added to _orphanedIndices
+    params ["_myStruct","_otherStruct","_navRoadHM"];
 
     private _myRoad = _myStruct#0;
     private _myConnections = _myStruct#1;
@@ -92,29 +79,25 @@ private _fnc_consumeStruct = {
     private _otherRoad = _otherStruct#0;
     private _otherName = str _otherRoad;
     private _otherConnections = _otherStruct#1;
-    private _otherConnectedStructs = _otherConnections apply {_navGridSimple #(_roadIndexNS getOrDefault [str _x,-1])};
+    private _otherConnectedStructs = _otherConnections apply {_navRoadHM get str _x};
 
     private _oldOtherConnections = +_otherConnections;
 
     {
         [_otherStruct,_x] call _fnc_disconnectStructs;
     } forEach _otherConnectedStructs;
-    if !((_navGridSimple #(_roadIndexNS getOrDefault [str _otherRoad,-1]) #1) isEqualTo A3A_NG_const_emptyArray) then {
+    if ((_navRoadHM get str _otherRoad) #1 isNotEqualTo A3A_NG_const_emptyArray) then {
         [1,"Tried to schedule deletion of non-orphan '"+_otherName+"' " + str (getPosWorld _otherRoad) + ".","fn_NG_simplify_junc"] call A3A_fnc_log;
         ["fn_NG_simplify_junc Error","Please check RPT."] call A3A_fnc_customHint;
     };
-    //if (_oldOtherConnections findIf (_otherRoad in (_navGridSimple #(_roadIndexNS getOrDefault [str _x,-1]) #1)) != -1) then {
-    //    [1,"Tried to schedule deletion of non-orphan that is connected from other roads'"+_otherName+"' " + str (getPosWorld _otherRoad) + ".","fn_NG_simplify_junc"] call A3A_fnc_log;
-    //    ["fn_NG_simplify_junc Error","Please check RPT."] call A3A_fnc_customHint;
-    //};
-    _orphanedIndices pushBack (_roadIndexNS getOrDefault [_otherName,-1]);
+    _navRoadHM deleteAt _otherName;
 
     {
         private _otherConnectedStruct = _x;
         private _otherConnectedRoad = _otherConnectedStruct#0;
 
         if !(_otherConnectedRoad in _myConnections) then {
-            if (_roadIndexNS getOrDefault [str _otherConnectedRoad,-1] in _orphanedIndices) then {
+            if !(str _otherConnectedRoad in _navRoadHM) then {
                 [1,"Tried to connect to orphan '"+str _otherConnectedRoad+"' " + str (getPosWorld _otherConnectedRoad) + ".","fn_NG_simplify_junc"] call A3A_fnc_log;
                 ["fn_NG_simplify_junc Error","Please check RPT."] call A3A_fnc_customHint;
             };
@@ -125,9 +108,8 @@ private _fnc_consumeStruct = {
 
 };
 
-
 private _fnc_selectCentreStruct = {
-    params ["_structs"]; // _otherStruct is consumed & added to _orphanedIndices
+    params ["_structs"];
 
     private _midPoint = [0,0];
     {
@@ -147,51 +129,36 @@ private _fnc_selectCentreStruct = {
     _centreStruct;
 };
 
-private _diag_totalSegments = count _navGridSimple;
+private _diag_totalSegments = count _navRoadHM;
 {
     if (_forEachIndex mod 100 == 0) then {
         _diag_step_sub = "Completion &lt;" + ((100*_forEachIndex /_diag_totalSegments) toFixed 1) + "% &gt; Processing segment &lt;" + (str _forEachIndex) + " / " + (str _diag_totalSegments) + "&gt;";;
         call _fnc_diag_render;
     };
 
-    private _myStruct = _x;
-    private _myRoad = _myStruct#0;
-    private _myConnections = _myStruct#1;
-    private _connectedJuncStructs = [];
-    if ((count _myConnections) > 2) then {
-        _connectedJuncStructs = _myConnections
-            select {_myRoad distance2D _x < _juncMergeDistance}                       // Only within small junction proximity
-            apply {_navGridSimple #(_roadIndexNS getOrDefault [str _x,-1])}     // Get their structs
-            select {count (_x#1) > 2};                                          // Only structs that are junctions
-    };
+    private _myStruct = _navRoadHM get _x;
+    if !(isNil {_myStruct}) then {  // If it is nil, it has been previously deleted.
+        private _myRoad = _myStruct#0;
+        private _myConnections = _myStruct#1;
+        private _connectedJuncStructs = [];
+        if ((count _myConnections) > 2) then {
+            _connectedJuncStructs = _myConnections
+                select {_myRoad distance2D _x < _juncMergeDistance}                         // Only within small junction proximity
+                apply {_navRoadHM get str _x}                                               // Get their structs
+                select {count (_x#1) > 2};                                                  // Only structs that are junctions
+        };
 
-    if ((count _connectedJuncStructs) != 0) then {  // At least one other junction to consume
-        _connectedJuncStructs pushBack _myStruct;
-        private _centreStruct = [_connectedJuncStructs] call _fnc_selectCentreStruct;
-        _connectedJuncStructs deleteAt (_connectedJuncStructs find _centreStruct);
+        if ((count _connectedJuncStructs) != 0) then {  // At least one other junction to consume
+            _connectedJuncStructs pushBack _myStruct;
+            private _centreStruct = [_connectedJuncStructs] call _fnc_selectCentreStruct;
+            _connectedJuncStructs deleteAt (_connectedJuncStructs find _centreStruct);
 
-        {
-            [_centreStruct,_x] call _fnc_disconnectStructs;
-            [_centreStruct,_x] call _fnc_consumeStruct;
-        } forEach _connectedJuncStructs;
+            {
+                [_centreStruct,_x] call _fnc_disconnectStructs;
+                [_centreStruct,_x,_navRoadHM] call _fnc_consumeStruct;
+            } forEach _connectedJuncStructs;
+        };
     };
-} forEach _navGridSimple;
+} forEach keys _navRoadHM;
 
-_diag_step_sub = "Cleaning orphans...";
-call _fnc_diag_render;
-//*
-{
-    if (_x == -1 || _x > count _navGridSimple) exitWith {
-        [1,"Tried to delete non-index -1.","fn_NG_simplify_junc"] call A3A_fnc_log;
-        ["fn_NG_simplify_junc Error","Tried to delete non-index -1. Please check RPT."] call A3A_fnc_customHint;
-        [];
-    };
-    if (count (_navGridSimple#_x#1) > 0) then {
-        [1,"Tried to delete non-orphan '"+str (_navGridSimple#_x#0)+"' " + str (getPosWorld (_navGridSimple#_x#0)) + ".","fn_NG_simplify_junc"] call A3A_fnc_log;
-        ["fn_NG_simplify_junc Error","Please check RPT."] call A3A_fnc_customHint;
-    };
-} forEach _orphanedIndices;//*/
-_orphanedIndices sort false; // The order of orphaned indices will be random.
-[_navGridSimple,_orphanedIndices] call Col_fnc_array_remIndices;
-
-_navGridSimple;
+_navRoadHM;
